add proper reorganisation for calypso.pc.Demo_CalypsoClassic
