+++ java/example/calypso/pc/Demo_CalypsoClassic/build.gradle
@@ -1,0 +1,31 @@
+plugins {
+    id 'java'
+}
+
+group 'org.eclipse.keyple'
+version '1.0-SNAPSHOT'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+    maven {
+        url 'https://repo.eclipse.org/service/local/repositories/maven_central/content'
+    }
+    //to import keyple snapshots
+    maven {url 'https://oss.sonatype.org/content/repositories/snapshots' }
+    //to import keyple releases
+    maven { url 'https://oss.sonatype.org/content/repositories/releases' }
+}
+
+dependencies {
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-core', version: '+'
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-calypso', version: '+'
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-pcsc', version: '+'
+    implementation group: 'org.eclipse.keyple', name: 'keyple-java-plugin-stub', version: '+'
+
+    implementation "org.slf4j:slf4j-simple:1.7.25"
+    implementation "org.slf4j:slf4j-ext:1.7.25"
+}
