+++ java/example/calypso/pc/Demo_CalypsoClassic/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'example.calypso.pc.Demo_CalypsoClassic'
+
