+++ java/example/calypso/pc/Demo_CalypsoClassic/src/main/java/org/eclipse/keyple/example/calypso/pc/Demo_CalypsoClassic/Demo_CalypsoClassic_Pcsc.java
@@ -9,7 +9,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
-package org.eclipse.keyple.example.calypso.pc;
+package org.eclipse.keyple.example.calypso.pc.Demo_CalypsoClassic;
 
 
 import org.eclipse.keyple.core.seproxy.SeProxyService;
@@ -18,8 +18,9 @@ import org.eclipse.keyple.core.seproxy.event.ObservableReader;
 import org.eclipse.keyple.core.seproxy.exception.KeypleBaseException;
 import org.eclipse.keyple.core.seproxy.exception.KeypleReaderNotFoundException;
 import org.eclipse.keyple.core.seproxy.protocol.SeCommonProtocols;
-import org.eclipse.keyple.example.calypso.pc.transaction.CalypsoClassicTransactionEngine;
-import org.eclipse.keyple.example.generic.pc.ReaderUtilities;
+import org.eclipse.keyple.example.common.PcscReadersSettings;
+import org.eclipse.keyple.example.common.calypso.pc.transaction.CalypsoClassicTransactionEngine;
+import org.eclipse.keyple.example.common.ReaderUtilities;
 import org.eclipse.keyple.plugin.pcsc.PcscPluginFactory;
 import org.eclipse.keyple.plugin.pcsc.PcscProtocolSetting;
 import org.eclipse.keyple.plugin.pcsc.PcscReaderSettings;
