+++ java/example/calypso/pc/Demo_CalypsoClassic/src/main/java/org/eclipse/keyple/example/common
@@ -1,0 +1,1 @@
+../../../../../../../../../../common/src/main/java/org/eclipse/keyple/example/common/
