+++ java/example/calypso/pc/Demo_CalypsoClassic/src/main/resources/config.properties
@@ -1,0 +1,1 @@
+../../../../../../common/src/main/resources/config.properties
